requirements.txt
gradio
fastapi
uvicorn
starlette
python-dotenv
google-generativeai
